+++ handler/src/main/java/io/netty/handler/internal/svm/ssl/JdkAlpnApplicationProtocolNegotiator$AlpnWrapperSubstitution.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 The Netty Project
+ * Copyright 2020 The Netty Project
  *
  * The Netty Project licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
